+++ src/main/java/roomescape/dao/ReservationDao.java
@@ -7,6 +7,7 @@ import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
+import roomescape.dao.mapper.ReservationRowMapper;
 import roomescape.domain.Reservation;
 import roomescape.domain.ReservationDate;
