+++ src/main/java/roomescape/dao/ReservationTimeDao.java
@@ -9,6 +9,7 @@ import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
+import roomescape.dao.mapper.ReservationTimeRowMapper;
 import roomescape.domain.ReservationTime;
 
 @Repository
